+++ internal/image/pathtree/pathtree.go
